factor components into modules
